Variable value editor change
Its not clear that you need to hit enter when entering variable values 
which means its easy to lose them by closing the modal.  I added an 
update and cancel button to make it clearer that you needed to do 
something, and give a way of backing out of a change and made it harder 
to close the modal mid change.